adds screens that will hold article index and about us info
