+++ App.js
@@ -1,25 +1,37 @@
 import { StatusBar } from 'expo-status-bar';
-import { useSelector } from 'react-redux'
 import React from 'react';
+import ArticlesIndex from './src/components/ArticlesIndex'
+import AboutUs from './src/components/AboutUs'
 import { StyleSheet, Text, View } from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
 
+
+const Stack = createStackNavigator()
 const App = () => {
-  const appName = useSelector(state => state.appName)
   return (
-    <View style={styles.container}>
-      <Text>Welcome to {appName}</Text>
-      <StatusBar style="auto" />
-    </View>
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen
+          name='ArticlesIndex'
+          component={ArticlesIndex}
+          options={
+            () => ({
+              headerStyle: {
+                backgroundColor: 'grey'
+              },
+              title: "Custom title"
+            })
+          }
+        />
+        <Stack.Screen
+          name='AboutUs'
+          component={AboutUs}
+        />
+      </Stack.Navigator>
+    </NavigationContainer>
+
   );
 }
 
 export default App
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
